requirements.txt
alembic==1.3.2
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cssmin==0.2.0
dash==1.7.0
dash-bootstrap-components==0.8.1
dash-core-components==1.6.0
dash-daq==0.3.1
dash-html-components==1.0.2
dash-renderer==1.2.2
dash-table==4.5.1
Flask==1.1.1
Flask-Assets==2.0
Flask-Compress==1.4.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
future==0.18.2
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==8.1.0
numpy==1.18.1
packaging==20.0
pandas==0.25.3
pathlib==1.0.1
plotly==4.4.1
plotly-flask-tutorial==0.0.1
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pyparsing==2.4.6
pytest==5.3.3
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
retrying==1.3.3
six==1.13.0
SQLAlchemy==1.3.12
urllib3==1.25.8
wcwidth==0.1.8
webassets==2.0
Werkzeug==0.16.0
